// vite-app/app/src/redux/interfaces.ts
export interface StoreState extends BaseStoreState, NarrativeState, ProfileState, OrgState { }

export interface UserAuthorization {
    realname: string;
    role?: Array<string>;
    token: string;
    username: string;
}

/**
 * Narrative 
 */
export interface NarrativeData {
    wsID: string;
    permission: string;
    name: string;
    last_saved: number;
    users: object;
    owner: string;
    narrative_detail: {
        creator: string;
    };
}

// used in reducer 
export interface NarrativeAction {
    type: string;
    payload: {
        narrativeList: Array<NarrativeData>;
        loading: boolean;
        isOwner: boolean;
    };
}

// need this for adding type to StoreState - see store.ts
export interface NarrativeState {
    narrativeState: {
        narrativeList: Array<NarrativeData>;
        loading: boolean;
        isOwner: boolean;
    };
}


/**
 * Orgs  
 */

// used in reducer 
// need this for adding type to StoreState - see store.ts
export interface OrgState {
    orgState: OrgList | OrgFetchStatus | OrgFetchError;
}

export interface OrgList {
    orgList: Array<OrgProp>,
    orgFetchStatus: AsyncFetchStatus.SUCCESS | AsyncFetchStatus.REFETCHING;
}

export interface OrgFetchError {
    orgError: Array<number | string>,
    orgFetchStatus: AsyncFetchStatus.ERROR;
}

export interface OrgFetchStatus {
    orgFetchStatus: AsyncFetchStatus.NONE | AsyncFetchStatus.FETCHING;
}
// ...
export interface ProfileState {
    profileView: ProfileView | ProfileFetchStatus | ErrorMessages;
}

export interface ErrorMessages {
    errorMessages: Array<number | string>;
    profileFetchStatus: AsyncFetchStatus.ERROR;
}

export interface ProfileFetchStatus {
    profileFetchStatus: AsyncFetchStatus.NONE | AsyncFetchStatus.FETCHING;
}

export interface ProfileView {
    userName: UserName,
    editEnable: boolean,
    profileUserdata: ProfileUserdata,
    gravatarHash: string,
    profileFetchStatus: AsyncFetchStatus.SUCCESS | AsyncFetchStatus.REFETCHING;
}

export interface ProfileUserdata {
    organization: string | null;
    department: string;
    city: string;
    state: string;
    postalCode: string;
    country: string;
    affiliations: Array<Affiliation>;
    researchStatement: string;
    jobTitle: string;
    jobTitleOther: string;
    researchInterests: Array<string>;
    researchInterestsOther: string | null;
    fundingSource: string;
    gravatarDefault: string;
    avatarOption: string;
}

export interface Affiliation {
    title: string;
    organization: string;
    started: number;
    ended: number | null;
}

// used in Profile View app
// TODO: please refactor, someone; having UserName as something
// other than a username string is unnecessarily confusing.
export interface UserName {
    name: string;
    userID: string;
}